Improved Vision for Reef detection
Reworked the reef vision pipeline to use an HSV Filter instead of an RGB Filter because it could detect the pipes much more accurately and consistently. Added commands to start the camera on different ports so that its possible to test the vision with the usb camera we will eventually use on the robot.